+++ src/modlunky2/sprites/base_classes/__init__.py
@@ -1,0 +1,5 @@
+from .base_sprite_loader import BaseSpriteLoader
+from .base_deco_sheet import AbstractDecoSheet
+from .base_biome import AbstractBiome
+
+__all__ = ["BaseSpriteLoader", "AbstractDecoSheet", "AbstractBiome"]
