+++ src/modlunky2/sprites/base_classes/types.py
