+++ src/modlunky2/sprites/deco_sheet.py
@@ -1,26 +1,4 @@
-from abc import abstractmethod
-from pathlib import Path
-from typing import Union
-
-from PIL import Image
-
-from .base_sprite_loader import BaseSpriteLoader
-
-number = Union[float, int]
-
-
-class AbstractDecoSheet(BaseSpriteLoader):
-    _sprite_sheet: Image.Image
-    _base_path: Path
-
-    @property
-    @abstractmethod
-    def biome_name(self) -> str:
-        raise NotImplementedError
-
-    @property
-    def _sprite_sheet_path(self):
-        return Path(f"Data/Textures/deco_{self.biome_name}.png")
+from .base_classes.base_deco_sheet import AbstractDecoSheet
 
 
 class CaveDecoSheet(AbstractDecoSheet):
