+++ src/modlunky2/sprites/floor_sheet.py
@@ -1,7 +1,7 @@
 from abc import abstractmethod
 from pathlib import Path
 
-from .base_sprite_loader import BaseSpriteLoader
+from .base_classes import BaseSpriteLoader
 
 
 class AbstractFloorSheet(BaseSpriteLoader):
